+++ apps/client/src/views/timer/Timer.scss
@@ -1,4 +1,4 @@
-@use '../../../theme/viewerDefs' as *;
+@use '../../theme/viewerDefs' as *;
 
 .stage-timer {
   margin: 0;
